requirements.txt
flask==1.1.2
gunicorn==20.0.4
pytest==5.4.1
jwt==1.0.0
sendgrid==6.2.2
jwt==1.0.0
Pyrebase==3.0.27
PyYAML==5.3.1